Write Data to a Text File
I figured out how to turn my die value int into a string and write it to a text file. I still need to figure out how to make an array that can count each value and hold them for the duration of the test, and then display those results.